+++ CHANGELOG.md
@@ -23,7 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added support for tax calculation where the values from customer_tax_class_ids is used - @resubaka (#3245)
 - Added loading product attributes (`entities.productListWithChildren.includeFields`) on category page - @andrzejewsky (#3220)
 - Added config to set Cache-Control header for static assets based on mime type - @phoenix-bjoern (#3268)
-- Added test:unit:watch with a workaround of a jest problem with template strings - @resubaka (#3351)
+- Added test:unit:watch with a workaround of a jest problem with template strings - @resubaka (#3351, #3354)
 
 ### Fixed